Many changes. moving things out into logical same classes. Moved somethings to their own file.
