+++ GoldBox.Classes/CombatType.cs
@@ -1,0 +1,8 @@
+﻿namespace GoldBox.Classes
+{
+    public enum CombatType
+    {
+        normal = 0,
+        duel = 1
+    }
+}
