+++ GoldBox.Classes/Control.cs
@@ -1,0 +1,11 @@
+﻿namespace GoldBox.Classes
+{
+    public class Control
+    {
+        public const byte PC_Base = 0;
+        public const byte PC_Mask = 0x7F;
+        public const byte NPC_Base = 0x80;
+        public const byte NPC_Berzerk = 0xB2;
+        public const byte PC_Berzerk = 0xB3;
+    }
+}
