+++ GoldBox.Classes/DamageOnSave.cs
@@ -1,0 +1,11 @@
+﻿namespace GoldBox.Classes
+{
+    public enum DamageOnSave
+    {
+        Normal = 0,
+        Zero = 1,
+        Half = 2,
+        Unknown_3 = 3,
+        Unknown_1E = 0x1e
+    }
+}
