+++ GoldBox.Classes/Delegates.cs
@@ -1,0 +1,6 @@
+﻿namespace GoldBox.Classes
+{
+    public delegate bool spellDelegate(QuickFight quick_fight, int spellId);
+    public delegate void spellDelegate2();
+    public delegate void affectDelegate(Effect arg_0, object affect, Player player);
+}
