+++ GoldBox.Classes/Effect.cs
@@ -1,0 +1,8 @@
+﻿namespace GoldBox.Classes
+{
+    public enum Effect
+    {
+        Add = 0,
+        Remove = 1
+    }
+}
