+++ GoldBox.Classes/GameState.cs
@@ -1,0 +1,15 @@
+﻿namespace GoldBox.Classes
+{
+    public enum GameState
+    {
+        StartGameMenu = 0,
+        Shop = 1,
+        Camping = 2,
+        WildernessMap = 3,
+        DungeonMap = 4,
+        Combat = 5,
+        AfterCombat = 6,
+        EndGame = 7
+    }
+
+}
