+++ GoldBox.Classes/ImportSource.cs
@@ -1,0 +1,9 @@
+﻿namespace GoldBox.Classes
+{
+    public enum ImportSource
+    {
+        Curse = 0,
+        Pool = 1,
+        Hillsfar = 2
+    }
+}
