+++ GoldBox.Classes/MenuColorSet.cs
@@ -1,0 +1,11 @@
+﻿namespace GoldBox.Classes
+{
+    public class MenuColorSet
+    {
+        public int highlight;
+        public int foreground;
+        public int prompt;
+
+        public MenuColorSet(int h, int f, int p) { highlight = h; foreground = f; prompt = p; }
+    }
+}
