+++ GoldBox.Classes/Player.cs
@@ -1,334 +1,8 @@
 using System;
 using System.Collections.Generic;
 
-
 namespace GoldBox.Classes
 {
-    public struct StatValue : IDataIO
-    {
-        int[, ,] raceSexMinMax;
-        int[] classMin;
-        int[] ageEffects;
-
-        public StatValue(int[, ,] _raceSexMinMax, int[] _classMin, int[] _ageEffects)
-        {
-            raceSexMinMax = _raceSexMinMax;
-            classMin = _classMin;
-            ageEffects = _ageEffects;
-            cur = full = 0;
-        }
-
-        public int cur;
-        public int full;
-
-        public void Load(int val)
-        {
-            full = cur = val;
-        }
-
-        public void Assign(StatValue sv)
-        {
-            full = sv.full;
-            cur = sv.cur;
-        }
-
-        public void Inc()
-        {
-            full += 1;
-            cur += 1;
-        }
-
-        public void Dec()
-        {
-            full -= 1;
-            cur -= 1;
-        }
-
-        public void EnforceRaceSexLimits(int race, int sex)
-        {
-            if( raceSexMinMax != null )
-            {
-                full = Math.Min(raceSexMinMax[race, 1, sex], full);
-                full = Math.Max(raceSexMinMax[race, 0, sex], full);
-            }
-            cur = full;
-        }
-
-        public void EnforceClassLimits(int _class)
-        {
-            if (classMin != null)
-            {
-                full = Math.Max(classMin[_class], full);
-            }
-            cur = full;
-        }
-
-        public void AgeEffects(int race, int age)
-        {
-            for (int i = 0; i < 5; i++)
-            {
-                if (Limits.RaceAgeBrackets[race, i] < age)
-                {
-                    full += ageEffects[i];
-                }
-            }
-        }
-
-        public void Write(byte[] data, int offset)
-        {
-            data[offset + 0] = (byte)cur;
-            data[offset + 1] = (byte)full;
-        }
-
-        public void Read(byte[] data, int offset)
-        {
-            // enforce values in valid range
-            cur = Math.Min((int)data[offset + 0], 25);
-            full = Math.Min((int)data[offset + 1], 25);
-        }
-
-        public override string ToString()
-        {
-            return string.Format("{0}/{1}", cur, full);
-        }
-    }
-
-    public class PlayerStats : IDataIO
-    {
-        public StatValue Str = new StatValue(Limits.StrRaceSexMinMax, Limits.StrClassMin, Limits.StrAgeEffect);
-        public StatValue Str00 = new StatValue(Limits.Str00RaceSexMinMax, Limits.Str00ClassMin, Limits.Str00AgeEffect);
-        public StatValue Con = new StatValue(Limits.ConRaceSexMinMax, Limits.ConClassMin, Limits.ConAgeEffect);
-        public StatValue Dex = new StatValue(Limits.DexRaceSexMinMax, Limits.DexClassMin, Limits.DexAgeEffect);
-        public StatValue Int = new StatValue(Limits.IntRaceSexMinMax, Limits.IntClassMin, Limits.IntAgeEffect);
-        public StatValue Wis = new StatValue(Limits.WisRaceSexMinMax, Limits.WisClassMin, Limits.WisAgeEffect);
-        public StatValue Cha = new StatValue(Limits.ChaRaceSexMinMax, Limits.ChaClassMin, Limits.ChaAgeEffect);
-
-
-        void IDataIO.Write(byte[] data, int offset)
-        {
-            Str.Write(data, offset + 0x00);
-            Int.Write(data, offset + 0x02);
-            Wis.Write(data, offset + 0x04);
-            Dex.Write(data, offset + 0x06);
-            Con.Write(data, offset + 0x08);
-            Cha.Write(data, offset + 0x0a);
-            Str00.Write(data, offset + 0x0c);
-        }
-
-        void IDataIO.Read(byte[] data, int offset)
-        {
-            Str.Read(data, offset + 0x00);
-            Int.Read(data, offset + 0x02);
-            Wis.Read(data, offset + 0x04);
-            Dex.Read(data, offset + 0x06);
-            Con.Read(data, offset + 0x08);
-            Cha.Read(data, offset + 0x0a);
-            Str00.Read(data, offset + 0x0c);
-        }
-
-        public void Assign(PlayerStats ps)
-        {
-            Str.Assign(ps.Str);
-            Int.Assign(ps.Int);
-            Wis.Assign(ps.Wis);
-            Dex.Assign(ps.Dex);
-            Con.Assign(ps.Con);
-            Cha.Assign(ps.Cha);
-            Str00.Assign(ps.Str00);
-        }
-
-		public void Dec(int idx)
-		{
-			switch (idx)
-			{
-				case 0: Str.Dec(); break;
-				case 1: Int.Dec(); break;
-				case 2: Wis.Dec(); break;
-				case 3: Dex.Dec(); break;
-				case 4: Con.Dec(); break;
-				case 5: Cha.Dec(); break;
-				default: throw new IndexOutOfRangeException(string.Format("idx {0} not in [0-5]", idx));
-			}
-		}
-
-		public void Inc(int idx)
-		{
-			switch (idx)
-			{
-				case 0: Str.Inc(); break;
-				case 1: Int.Inc(); break;
-				case 2: Wis.Inc(); break;
-				case 3: Dex.Inc(); break;
-				case 4: Con.Inc(); break;
-				case 5: Cha.Inc(); break;
-				default: throw new IndexOutOfRangeException(string.Format("idx {0} not in [0-5]", idx));
-			}
-		}
-
-        public StatValue this[int idx]
-        {
-            get
-            {
-                switch (idx)
-                {
-                    case 0: return Str;
-                    case 1: return Int;
-                    case 2: return Wis;
-                    case 3: return Dex;
-                    case 4: return Con;
-                    case 5: return Cha;
-                    default: throw new IndexOutOfRangeException(string.Format("idx {0} not in [0-5]", idx));
-                }
-            }
-        }
-
-        public override string ToString()
-        {
-            return string.Format("S:{0} ({6}),I:{1},W:{2},C:{3},D:{4},Ch:{5}", Str, Int, Wis, Con, Dex, Cha, Str00);
-        }
-    }
-
-    public struct ClassLevels
-    {
-
-    }
-
-    public class ActiveItems
-    {
-        const int ItemSlots = 13;
-        Item[] itemArray = new Item[ItemSlots]; // 0x151[]
-
-        public byte PrimaryWeaponHandCount()
-        {
-            if (primaryWeapon == null) return 0;
-
-            return primaryWeapon.HandsCount();
-        }
-
-        public byte SecondaryWeaponHandCount()
-        {
-            if (secondaryWeapon == null) return 0;
-
-            return secondaryWeapon.HandsCount();
-        }     
-
-        public Item primaryWeapon // field_151
-        {// 0x151
-            get { return itemArray[0]; }
-            set { itemArray[0] = value; }
-        }
-        public Item secondaryWeapon // field_155
-        { // 0x155 
-            get { return itemArray[1]; }
-            set { itemArray[1] = value; }
-        }
-
-        public Item armor // field_159
-        { // 0x159 
-            get { return itemArray[2]; }
-            set { itemArray[2] = value; }
-        }
-
-        public Item field_15D
-        { // 0x15D
-            get { return itemArray[3]; }
-            set { itemArray[3] = value; }
-        }
-        public Item field_161
-        { // 0x161
-            get { return itemArray[4]; }
-            set { itemArray[4] = value; }
-        }
-        public Item field_165
-        { // 0x165
-            get { return itemArray[5]; }
-            set { itemArray[5] = value; }
-        }
-        public Item field_169
-        { // 0x169 
-            get { return itemArray[6]; }
-            set { itemArray[6] = value; }
-        }
-        public Item field_16D
-        { // 0x16D
-            get { return itemArray[7]; }
-            set { itemArray[7] = value; }
-        }
-        public Item field_171
-        { // 0x171
-            get { return itemArray[8]; }
-            set { itemArray[8] = value; }
-        }
-        public Item Item_ptr_01
-        { // 0x175
-            get { return itemArray[9]; }
-            set { itemArray[9] = value; }
-        }
-        public Item Item_ptr_02
-        {// 0x179
-            get { return itemArray[10]; }
-            set { itemArray[10] = value; }
-        }
-        public Item arrows
-        {// 0x17d
-            get { return itemArray[11]; }
-            set { itemArray[11] = value; }
-        }
-        public Item quarrels
-        { // 0x181
-            get { return itemArray[12]; }
-            set { itemArray[12] = value; }
-        }
-
-        public Item this[ItemSlot slot]
-        {
-            get
-            {
-                return itemArray[(int)slot];
-            }
-            set
-            {
-                itemArray[(int)slot] = value;
-            }
-        }
-
-        public void Reset()
-        {
-            for (int slot = 0; slot < ItemSlots; slot++)
-            {
-                itemArray[slot] = null;
-            }
-        }
-
-        public void UndreadyAll(int classFlags)
-        {
-            for (int item_slot = 0; item_slot < ItemSlots; item_slot++)
-            {
-                if (itemArray[item_slot] != null &&
-                    (gbl.ItemDataTable[itemArray[item_slot].type].classFlags & classFlags) == 0 &&
-                    itemArray[item_slot].cursed == false)
-                {
-                    itemArray[item_slot].readied = false;
-                }
-            }
-        }
-    }
-
-
-    public class Control
-    {
-        public const byte PC_Base = 0;
-        public const byte PC_Mask = 0x7F;
-        public const byte NPC_Base = 0x80;
-        public const byte NPC_Berzerk = 0xB2;
-        public const byte PC_Berzerk = 0xB3;
-    }
-
-
-
-    /// <summary>
-    /// Summary description for Player.
-    /// </summary>
     public class Player
     {
         [DataOffset(0x00, DataType.PString, 15)]
