+++ GoldBox.Classes/QuickFight.cs
@@ -1,0 +1,8 @@
+﻿namespace GoldBox.Classes
+{
+    public enum QuickFight
+    {
+        False = 0,
+        True = 1
+    }
+}
