+++ GoldBox.Classes/Sound.cs
@@ -1,0 +1,23 @@
+﻿namespace GoldBox.Classes
+{
+    public enum Sound
+    {
+        sound_FF = -1,
+        sound_0 = 0,
+        sound_1 = 1,
+        sound_2 = 2,
+        sound_3 = 3,
+        sound_4 = 4,
+        sound_5 = 5,
+        sound_6 = 6,
+        sound_attackHeld = 7,
+        sound_8 = 8,
+        sound_9 = 9,
+        sound_a = 0xa,
+        sound_b = 0xb,
+        sound_c = 0xc,
+        sound_d = 0xd,
+        sound_e = 0xe,
+        sound_f = 0xf
+    }
+}
