+++ GoldBox.Classes/SoundType.cs
@@ -1,0 +1,8 @@
+﻿namespace GoldBox.Classes
+{
+    public enum SoundType
+    {
+        PC,
+        None
+    }
+}
