+++ GoldBox.Classes/TextRegion.cs
@@ -1,0 +1,9 @@
+﻿namespace GoldBox.Classes
+{
+    public enum TextRegion
+    {
+        NormalBottom,
+        Normal2,
+        CombatSummary,
+    }
+}
