+++ GoldBox.Engine/ovr002.cs
@@ -6,6 +6,36 @@ namespace GoldBox.Engine
 {
     class ovr002
     {
+        internal static void title_screen()
+        {
+            DaxBlock dax_ptr;
+
+            dax_ptr = seg040.LoadDax(0, 0, 1, "Title");
+            seg040.draw_picture(dax_ptr, 0, 0, 0);
+
+            delay_or_key(5);
+
+            dax_ptr = seg040.LoadDax(0, 0, 2, "Title");
+            seg040.draw_picture(dax_ptr, 0, 0, 0);
+
+            dax_ptr = seg040.LoadDax(0, 0, 3, "Title");
+            seg040.draw_picture(dax_ptr, 0x0b, 6, 0);
+            delay_or_key(10);
+
+            dax_ptr = seg040.LoadDax(0, 0, 4, "Title");
+
+            seg044.PlaySound(Sound.sound_d);
+
+            seg040.draw_picture(dax_ptr, 0x0b, 0, 0);
+            delay_or_key(10);
+
+            seg041.ClearScreen();
+            credits();
+            delay_or_key(10);
+
+            seg041.ClearScreen();
+        }
+
         static void delay_or_key(int seconds)
         {
             seg043.clear_keyboard();
@@ -20,7 +50,6 @@ namespace GoldBox.Engine
 
             seg043.clear_keyboard();
         }
-
 
         static void credits()
         {
@@ -64,37 +93,6 @@ namespace GoldBox.Engine
             seg041.displayString("robert daly", 0x0, 0x0B, 0x16, 0x16);
 
             Display.UpdateStart();
-        }
-
-
-        internal static void title_screen()
-        {
-            DaxBlock dax_ptr;
-
-            dax_ptr = seg040.LoadDax(0, 0, 1, "Title");
-            seg040.draw_picture(dax_ptr, 0, 0, 0);
-
-            delay_or_key(5);
-
-            dax_ptr = seg040.LoadDax(0, 0, 2, "Title");
-            seg040.draw_picture(dax_ptr, 0, 0, 0);
-
-            dax_ptr = seg040.LoadDax(0, 0, 3, "Title");
-            seg040.draw_picture(dax_ptr, 0x0b, 6, 0);
-            delay_or_key(10);
-
-            dax_ptr = seg040.LoadDax(0, 0, 4, "Title");
-
-            seg044.PlaySound(Sound.sound_d);
-
-            seg040.draw_picture(dax_ptr, 0x0b, 0, 0);
-            delay_or_key(10);
-
-            seg041.ClearScreen();
-            credits();
-            delay_or_key(10);
-
-            seg041.ClearScreen();
         }
     }
 }
