+++ GoldBox.Engine/seg051.cs
@@ -5,8 +5,6 @@ namespace GoldBox.Engine
     class seg051
     {
         static System.Random random_number;
-        /*static int random_number;*/
-
 
         internal static string Copy(int CopyLen, int StartAt, string InString)
         {
@@ -58,34 +56,6 @@ namespace GoldBox.Engine
         internal static void Randomize()
         {
             random_number = new System.Random(unchecked((int)System.DateTime.Now.Ticks));
-        }
-
-
-        internal static void Reset(File arg_4)
-        {
-            arg_4.stream.Seek(0, System.IO.SeekOrigin.Begin);
-        }
-
-        internal static void Rewrite(File arg_2)
-        {
-            arg_2.stream.SetLength(0);
-        }
-
-
-        internal static void Close(File arg_0)
-        {
-            arg_0.stream.Close();
-        }
-
-        internal static int BlockRead(int count, byte[] data, File file)
-        {
-            return file.stream.Read(data, 0, count);
-        }
-
-
-        internal static void BlockWrite(int arg_4, byte[] arg_6, File arg_A)
-        {
-            arg_A.stream.Write(arg_6, 0, arg_4);
         }
 
         internal static void FillChar(byte fill_byte, int buffer_size, byte[] buffer)
