+++ coab.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoldBox.Engine", "GoldBox.Engine\GoldBox.Engine.csproj", "{361087AE-2FD8-4BEF-8218-05A553ED7A49}"
 EndProject
@@ -17,6 +17,10 @@ Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "Installer", "Installer\Inst
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoldBox.Logging", "GoldBox.Logging\GoldBox.Logging.csproj", "{BEC15EB8-3A21-49FF-ABC0-7465927E80B8}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{488B5DF3-7EBD-463D-AAE4-22238C5B7532}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoldBox.Engine.Tests", "GoldBox.Engine.Tests\GoldBox.Engine.Tests.csproj", "{D3606859-8487-4659-9B18-A8D24E07BE4E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -62,8 +66,19 @@ Global
 		{BEC15EB8-3A21-49FF-ABC0-7465927E80B8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BEC15EB8-3A21-49FF-ABC0-7465927E80B8}.Release|x86.ActiveCfg = Release|x86
 		{BEC15EB8-3A21-49FF-ABC0-7465927E80B8}.Release|x86.Build.0 = Release|x86
+		{D3606859-8487-4659-9B18-A8D24E07BE4E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D3606859-8487-4659-9B18-A8D24E07BE4E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D3606859-8487-4659-9B18-A8D24E07BE4E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D3606859-8487-4659-9B18-A8D24E07BE4E}.Debug|x86.Build.0 = Debug|Any CPU
+		{D3606859-8487-4659-9B18-A8D24E07BE4E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D3606859-8487-4659-9B18-A8D24E07BE4E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D3606859-8487-4659-9B18-A8D24E07BE4E}.Release|x86.ActiveCfg = Release|Any CPU
+		{D3606859-8487-4659-9B18-A8D24E07BE4E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{D3606859-8487-4659-9B18-A8D24E07BE4E} = {488B5DF3-7EBD-463D-AAE4-22238C5B7532}
 	EndGlobalSection
 EndGlobal
